group details test
